[core] Refactor core crate.
* Move remaining functionality to embed module.
* Runtime::handle -> context_handle
* Free functions are thin wrappers around RuntimeContext.
* moxie::embed -> moxie::runtime
* Remove built-in executor.
  * It was only used for testing, replace with futures LocalPool.
