+++ benches/core.rs
@@ -3,8 +3,8 @@ extern crate criterion;
 
 use criterion::{Criterion, ParameterizedBenchmark};
 use moxie::{
-    embed::{Revision, Runtime},
     once,
+    runtime::{Revision, RunLoop},
 };
 use std::rc::Rc;
 
@@ -12,22 +12,21 @@ criterion::criterion_group!(runtime, once_from_store, run_empty, run_repeated);
 criterion::criterion_main!(runtime);
 
 fn once_from_store(c: &mut Criterion) {
-    let mut rt = Runtime::new();
-    let run = || once(|| Rc::new(vec![0; 1_000_000]));
-    rt.run_once(run);
-    c.bench_function("1mb vec cached", |b| b.iter(|| rt.run_once(run)));
+    let mut rt = RunLoop::new(|| once(|| Rc::new(vec![0; 1_000_000])));
+    rt.run_once();
+    c.bench_function("1mb vec cached", |b| b.iter(|| rt.run_once()));
 }
 
 fn run_empty(c: &mut Criterion) {
-    let mut rt = Runtime::new();
-    c.bench_function("run_empty", |b| b.iter(|| rt.run_once(Revision::current)));
+    let mut rt = RunLoop::new(Revision::current);
+    c.bench_function("run_empty", |b| b.iter(|| rt.run_once()));
 }
 
 fn run_n_times_empty(b: &mut criterion::Bencher, n: &usize) {
-    let mut rt = Runtime::new();
+    let mut rt = RunLoop::new(Revision::current);
     b.iter(|| {
         for _ in 0..*n {
-            rt.run_once(Revision::current);
+            rt.run_once();
         }
     });
 }
